Rename Utilities.Collection namespace to plural
