+++ Obsidian/ChunkData/BiomeContainer.cs
@@ -1,5 +1,5 @@
 ﻿using Obsidian.Net;
-using Obsidian.Utilities.Collection;
+using Obsidian.Utilities.Collections;
 
 namespace Obsidian.ChunkData;
