+++ Obsidian/Utilities/Collections/DataArray.cs
@@ -1,6 +1,6 @@
 ﻿using System.Diagnostics;
 
-namespace Obsidian.Utilities.Collection;
+namespace Obsidian.Utilities.Collections;
 
 // Utility storage that can store multiple elements in 64-bit storage entries
 // This type's storage must correspond to minecraft's compacted data array
